pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "arize-phoenix>=12.29.0",
    "bcrypt>=5.0.0",
    "datasets>=4.4.2",
    "deepeval>=3.7.9",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=4.1.1",
    "flask-socketio>=5.6.0",
    "hiredis>=3.3.0",
    "joblib>=1.5.3",
    "openai>=2.15.0",
    "openinference-instrumentation-openai>=0.1.41",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.2",
    "redis>=7.1.0",
    "sentence-transformers>=5.2.0",
    "weaviate-client>=4.19.2",
    "werkzeug>=3.1.5",
]
